modified:   backend/cron/cron.js 	modified:   backend/index.js 	modified:   backend/routes/routes.js 	modified:   backend/vercel.json 	modified:   vercel.json

+++ backend/cron/cron.js
@@ -1,36 +1,38 @@
-const cron = require('node-cron');
-const { Op } = require('sequelize'); // Import Op from sequelize
-const Message = require('../models/message');
-const { sequelize } = require('../config/db');
+import { Op } from 'sequelize'; // Import Op from Sequelize
+import Message from '../models/message'; // Adjust the import based on your folder structure
+import { sequelize } from '../config/db'; // Adjust the import based on your folder structure
 
-module.exports = () => {
-  cron.schedule('0 * * * *', async () => {
-    console.log("Running scheduled job to delete old messages...");
+export default async function handler(req, res) {
+  // Check authorization
+  if (req.headers.authorization !== `Bearer ${process.env.CRON_SECRET}`) {
+    return res.status(401).end('Unauthorized');
+  }
 
-    const twoDaysAgo = new Date();
-    twoDaysAgo.setHours(twoDaysAgo.getHours() - 48);
+  // Define the time threshold for deleting messages
+  const twoDaysAgo = new Date();
+  twoDaysAgo.setHours(twoDaysAgo.getHours() - 48);
 
-    try {
-      // Check if the database connection is still valid
-      await sequelize.authenticate();
-
-      const result = await Message.destroy({
-        where: {
-          timestamp: {
-            [Op.lt]: twoDaysAgo, // Use Op.lt for querying
-          },
+  try {
+    await sequelize.authenticate();
+    const result = await Message.destroy({
+      where: {
+        timestamp: {
+          [Op.lt]: twoDaysAgo,
         },
-      });
+      },
+    });
 
-      console.log(`${result} old messages deleted.`);
-      console.log(`Messages deleted before: ${twoDaysAgo}`);
-
-    } catch (error) {
-      if (error.name === 'SequelizeDatabaseError') {
-        console.error('Database error deleting old messages:', error);
-      } else {
-        console.error('Error deleting old messages:', error);
-      }
+    console.log(`${result} old messages deleted.`);
+    console.log(`Messages deleted before: ${twoDaysAgo}`);
+    
+    res.status(200).end(`${result} old messages deleted.`);
+  } catch (error) {
+    if (error.name === 'SequelizeDatabaseError') {
+      console.error('Database error deleting old messages:', error);
+      return res.status(500).end('Database error');
+    } else {
+      console.error('Error deleting old messages:', error);
+      return res.status(500).end('Internal server error');
     }
-  });
-};
+  }
+}
